requirements.txt
streamlit
langchain
langgraph
langchain-groq
langchain-community
sqlalchemy
pandas
matplotlib
pillow
ipython
tabulate